Extract Action base class to eliminate SuccessAction/FailureAction duplication
Move shared fields name, type, workflowId, stepId, criteria and their
getters/setters/addCriteria into a common Action base class. Both
SuccessAction and FailureAction now extend it, retaining only their
own fluent builder methods and FailureAction-specific retry fields.
