+++ src/main/java/com/apiflows/model/FailureAction.java
@@ -1,49 +1,9 @@
 package com.apiflows.model;
 
-import java.util.ArrayList;
-import java.util.List;
+public class FailureAction extends Action {
 
-public class FailureAction {
-
-    private String name;
-    private String type;
-    private String workflowId;
-    private String stepId;
     private Double retryAfter;
     private Integer retryLimit;
-    private List<Criterion> criteria;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getWorkflowId() {
-        return workflowId;
-    }
-
-    public void setWorkflowId(String workflowId) {
-        this.workflowId = workflowId;
-    }
-
-    public String getStepId() {
-        return stepId;
-    }
-
-    public void setStepId(String stepId) {
-        this.stepId = stepId;
-    }
 
     public Double getRetryAfter() {
         return retryAfter;
@@ -61,31 +21,23 @@ public class FailureAction {
         this.retryLimit = retryLimit;
     }
 
-    public List<Criterion> getCriteria() {
-        return criteria;
-    }
-
-    public void setCriteria(List<Criterion> criteria) {
-        this.criteria = criteria;
-    }
-
     public FailureAction name(String name) {
-        this.name = name;
+        setName(name);
         return this;
     }
 
     public FailureAction type(String type) {
-        this.type = type;
+        setType(type);
         return this;
     }
 
     public FailureAction stepId(String stepId) {
-        this.stepId = stepId;
+        setStepId(stepId);
         return this;
     }
 
     public FailureAction workflowId(String workflowId) {
-        this.workflowId = workflowId;
+        setWorkflowId(workflowId);
         return this;
     }
 
@@ -97,12 +49,5 @@ public class FailureAction {
     public FailureAction retryLimit(Integer retryLimit) {
         this.retryLimit = retryLimit;
         return this;
-    }
-
-    public void addCriteria(Criterion criterion) {
-        if(this.criteria == null) {
-            this.criteria = new ArrayList<>();
-        }
-        this.criteria.add(criterion);
     }
 }
