+++ src/main/java/com/apiflows/model/SuccessAction.java
@@ -1,80 +1,24 @@
 package com.apiflows.model;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class SuccessAction {
-
-    private String name;
-    private String type;
-    private String workflowId;
-    private String stepId;
-    private List<Criterion> criteria;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getWorkflowId() {
-        return workflowId;
-    }
-
-    public void setWorkflowId(String workflowId) {
-        this.workflowId = workflowId;
-    }
-
-    public String getStepId() {
-        return stepId;
-    }
-
-    public void setStepId(String stepId) {
-        this.stepId = stepId;
-    }
-
-    public List<Criterion> getCriteria() {
-        return criteria;
-    }
-
-    public void setCriteria(List<Criterion> criteria) {
-        this.criteria = criteria;
-    }
-
-    public void addCriteria(Criterion criterion) {
-        if(this.criteria == null) {
-            this.criteria = new ArrayList<>();
-        }
-        this.criteria.add(criterion);
-    }
+public class SuccessAction extends Action {
 
     public SuccessAction name(String name) {
-        this.name = name;
+        setName(name);
         return this;
     }
 
     public SuccessAction type(String type) {
-        this.type = type;
+        setType(type);
         return this;
     }
 
     public SuccessAction workflowId(String workflowId) {
-        this.workflowId = workflowId;
+        setWorkflowId(workflowId);
         return this;
     }
 
     public SuccessAction stepId(String stepId) {
-        this.stepId = stepId;
+        setStepId(stepId);
         return this;
     }
 }
